agg_by {SERNAC}	R Documentation

Agregar/resumir por variable(s)

Usage

agg_by(
  df,
  bycols = c("proveedor_rut", "proveedor_mercado_nombre"),
  timep = "mensual",
  prep = FALSE,
  prepIgnore = NULL,
  fillN = TRUE
)

Arguments

df
data.table data.frame

bycols
columnas que agruparán

timep
tiempo para hacer la agrupación; mensual o semanal

prep
entrega una tabla lista para ser apilada con otras, sin perder información. Sólo si (setdiff(bycols, prepIgnore)) tiene largo igual a 1.} \item{prepIgnore}{columna de agrupación que ignorar al preparar los datos} \item{fillN}{cambia NAs por 0 (sólo en N, no en diff).} } { un data.frame de data.table ordenado por tiempo, para llegar y aplicar una serie temporal (filtrando previamente). } { Agrupa reclamos por fecha, de forma mensual o semanal. Retorno los tiempos de respuesta entre reclamos, y el número de reclamos contabilizados para esa agrupación temporal, con diferencias lagueadas definidas por el usuario. También agrupa por variable, dependiendo de lo que defina el usuario. } { 1+1 }

[Package SERNAC version 0.1.0 Index]