computar_reclamos {SERNAC}	R Documentation

Procesar reclamos

Description

Procesar reclamos

Usage

computar_reclamos(
  mdatai,
  byCategoria = c("cierre_corto", "proveedor_mercado_nombre",
    "proveedor_mercado_categoria_nombre", "proveedor_rut", "cut_region"),
  byClase = c(industria = "proveedor_mercado_categoria_nombre", proveedor =
    "proveedor_rut", producto = "mercado_tipo_producto_nombre"),
  lags = 1:6,
  cuantiles = c(0, 0.25, 0.5, 0.75, 1),
  IQR = c(0.25, 0.75),
  coef = 1.5,
  n_min = 50,
  rango_fechas = c(NULL, NULL)
)

Arguments

mdatai
data.frame con la base de datos.

byCategoria
vector de strings con las variables consideradas como categorías.

byClase
vector de strings con nombres ('setNames'), que incluya las variables de clase (categorización). El orden es importante.

lags
vector númerico

cuantiles
vector númerico con los cuantiles deseados (valor entre 0 y 1).

IQR
vector con el rango intercuantil de la forma c(min, max). Ambos valores deben estar en 'cuantiles'.

coef
coeficiente de dispersión (de castigo), usualmente 1.5.

n_min
número mínimo de observaciones

rango_fechas
vector de la forma 'c(min, max)' que incluye las fechas inferior y superior a considerar. Puede ser en formato fecha o como string en formato YYYY-mm-dd.

[Package SERNAC version 0.1.0 Index]