homologar_db {SERNAC}	R Documentation

Homologa datos de entrada al formato apropiado

Description

Permite homologar nombres de columnas, comunas, motivos legales; también selecciona columnas de interés, y añade datos del SII. La tabla queda lista para ser añadida a la tabla general de datos. La función trata de ser lo más inteligente posible, pero es necesario mantener los formatos apropiados.

Usage

homologar_db(
  db,
  diccionario_columnas = NULL,
  codigos_comunales = NULL,
  arbol_motivo_legal = NULL,
  datos_sii = NULL,
  verbose = TRUE,
  full_output = TRUE
)

Arguments

db
ruta al archivo, en excel (formato especial flexible).

diccionario_columnas
ruta al archivo de diccionario de columnas en excel (formato especial estricto).

codigos_comunales
ruta al archivo de codigos y nombres comunales en excel (formato especial estricto). Si 'skip', se salta este paso.

arbol_motivo_legal
ruta al archivo de equivalencias de motivos legales en excel (formato especial estricto).

datos_sii
ruta al archivo de datos del SII, en csv (formato especial estricto).

verbose
si TRUE, imprime estados de avance.

full_output
si TRUE, entrega todas las tablas usadas. Si FALSE, entrega solo la tabla principal modificada.

[Package SERNAC version 0.1.0 Index]