seleccionar_ranking {SERNAC}	R Documentation

Seleccionar por ranking

Description

Seleccionar por ranking

Usage

seleccionar_ranking(
  reclamos,
  byCategoria = byCategoria,
  topn,
  nmin,
  reporte = TRUE,
  ...
)

Arguments

reclamos
base de datos de reclamos

byCategoria
vector de strings con las variables consideradas como categorías.

topn
número de posiciones del ranking a incluir. Se pueden retornar más o menos valores, dependiendo de si existen empates o el número final de registros.

nmin
número mínimo de datos requeridos para estar presente en la selección.

reporte
si es TRUE, imprime un reporte resumen.

...
argumentos extras pasados a rankear.

Examples

1+1

[Package SERNAC version 0.1.0 Index]